refactor(plyphon): rename ui module to egui, one submodule per node
Rename the `ui` module to `egui` so a future alternate UI backend
(e.g. bevy-ui) can land as a sibling module behind its own feature.
The only naming collision with the extern egui crate is at the crate
root, where the re-export disambiguates with `self::`. Submodule and
expression paths are unaffected, since crate-root items are not in
their scope.

Split the single node.rs into one submodule per node
(egui/node/{bus,lag,out,pack,scope_out,sin_osc,unpack}.rs), mirroring
the crate::node layout so each node's behaviour and its egui surface
sit at corresponding paths.
