+++ crates/gantz_plyphon/src/egui/node/pack.rs
@@ -1,0 +1,50 @@
+//! `~pack`'s egui implementation.
+
+use crate::egui::param::value_row;
+use crate::node::Pack;
+use gantz_egui::{
+    InspectorRowsResponse, NodeCtx, NodeUi, NodeUiResponse, Registry, SocketDoc, SocketKind,
+};
+
+impl NodeUi for Pack {
+    fn name(&self, _: &dyn Registry) -> &str {
+        "~pack"
+    }
+
+    fn description(&self) -> Option<&'static str> {
+        Some("Concatenate signals into one multichannel signal (packed, not mixed)")
+    }
+
+    fn ui(&mut self, _ctx: NodeCtx, uictx: egui_graph::NodeCtx) -> NodeUiResponse {
+        let framed =
+            uictx.framed(|ui, _sockets| ui.add(egui::Label::new("~pack").selectable(false)));
+        NodeUiResponse::new(framed)
+    }
+
+    fn inspector_rows(
+        &mut self,
+        _ctx: &mut NodeCtx,
+        body: &mut egui_extras::TableBody,
+    ) -> InspectorRowsResponse {
+        let mut resp = InspectorRowsResponse::default();
+        // Input count (structural: it changes the node's sockets -> respawn).
+        let mut count = self.count();
+        let dv = egui::DragValue::new(&mut count).range(1..=64).speed(1.0);
+        if value_row(body, "count", dv) {
+            self.set_count(count);
+            resp.mark_changed();
+        }
+        resp
+    }
+
+    fn socket_doc(&self, _: &dyn Registry, kind: SocketKind, ix: usize) -> Option<SocketDoc> {
+        match kind {
+            SocketKind::Input => Some(SocketDoc::ty("signal").with_description(format!(
+                "signal {ix} to pack (any channel width; silence when unconnected)"
+            ))),
+            SocketKind::Output => Some(SocketDoc::ty("signal").with_description(
+                "every input's channels concatenated (width = the sum of input widths)",
+            )),
+        }
+    }
+}
