+++ crates/gantz_plyphon/src/node/mod.rs
@@ -3,7 +3,7 @@
 //! Each node implements [`gantz_core::Node`] (a placeholder, since DSP nodes are
 //! inert in the Steel world), [`NodeDsp`](crate::NodeDsp) (the audio behaviour)
 //! and [`ToNodeDsp`](crate::ToNodeDsp) (discovery). Their `gantz_egui::NodeUi`
-//! impls live in `crate::ui` (`egui` feature). The `~` keyword-name prefix
+//! impls live in `crate::egui` (`egui` feature). The `~` keyword-name prefix
 //! marks them as dsp nodes.
 
 pub use bus::Bus;
